instance-manager/requirements.txt
jsonschema
PyYAML
gunicorn
gevent
python-dotenv
kubernetes
requests
Flask
Flask-Cors
psutil
